Fix model download failures: FCOO retry/fallback, CMEMS memory
- FCOO: retry download up to 3x with HTTP Range resume on ChunkedEncodingError;
  fall back to most recent cached file if all attempts fail (handles today's
  server drops at ~5% into the 1.9GB file)
- CMEMS: limit open_dataset to a 8-day window (now-1d to now+7d) via
  start/end_datetime params; subsample spatial grid 2x (lat/lon every 2nd
  point); keep float32 in build_interpolators — reduces peak memory from
  ~24GB to ~2.5GB, fitting within GitHub Actions' 7GB RAM limit
- run.py: pass fetch_start/fetch_end to download_model_data in
  _extend_trajectories

+++ src/simulate.py
@@ -102,12 +102,14 @@ def build_interpolators(model_data: xr.Dataset) -> tuple[Callable, Callable]:
     lat   = model_data["lat"].values.astype(np.float64)
     lon   = model_data["lon"].values.astype(np.float64)
 
-    u_arr = model_data["u"].values.astype(np.float64)  # (time, depth, lat, lon)
-    v_arr = model_data["v"].values.astype(np.float64)
+    # float32 halves the memory vs float64; position errors at km scale
+    # don't require mm/s current precision.
+    u_arr = model_data["u"].values.astype(np.float32)  # (time, depth, lat, lon)
+    v_arr = model_data["v"].values.astype(np.float32)
 
     # Replace NaN (land/mask) with 0 so interpolation doesn't propagate NaNs
-    u_arr = np.where(np.isnan(u_arr), 0.0, u_arr)
-    v_arr = np.where(np.isnan(v_arr), 0.0, v_arr)
+    u_arr = np.where(np.isnan(u_arr), np.float32(0.0), u_arr)
+    v_arr = np.where(np.isnan(v_arr), np.float32(0.0), v_arr)
 
     # Ensure coordinate arrays are strictly monotone (required by RGI)
     if depth[0] > depth[-1]:
